Fix class label retrieval by casting class_ids to size_t in output display

+++ src/main.cpp
@@ -93,7 +93,8 @@ int main(int argc, char *argv[]) {
             std::cout << (use_segmentation ? "Instance " : "Detection ") << i << ":" << std::endl;
             std::cout << "  Box: [" << boxes[i][0] << ", " << boxes[i][1] << ", " << boxes[i][2] << ", " << boxes[i][3]
                       << "]" << std::endl;
-            std::cout << "  Class: " << inference.get_coco_labels()[class_ids[i]] << " (Score: " << scores[i] << ")"
+            std::cout << "  Class: " << inference.get_coco_labels()[static_cast<size_t>(class_ids[i])]
+                      << " (Score: " << scores[i] << ")"
                       << std::endl;
             if (use_segmentation && i < masks.size()) {
                 const int mask_pixels = cv::countNonZero(masks[i]);
